+++ example/a.ts
@@ -1,14 +1,30 @@
-import { Elysia, t } from '../src'
-import { req } from '../test/utils'
+import { Elysia, t, ValidationError } from '../src'
+import { post, req } from '../test/utils'
 
+let err: Error | undefined
 
 const app = new Elysia()
-			// @ts-ignore
-			.onRequest(({ qi }) => {
-				// queryIndex = qi
-				console.log("A", qi)
-			})
-			.get('/', () => 'ok')
-			.listen(0)
+	.post('/', ({ body }) => body, {
+		headers: t.Object({
+			year: t.Numeric({ minimum: 1900, maximum: 2160 })
+		}),
+		transform({ headers }) {
+			console.log(headers)
+		},
+		error({ code, error }) {
+			console.log(code)
+			switch (code) {
+				case 'VALIDATION':
+					err = error
+			}
+		}
+	})
+	.listen(3000)
 
-		await fetch(`http://localhost:${app.server?.port}`)
+await app.handle(
+	req('/', {
+		headers: {
+			year: '3000'
+		}
+	})
+)