+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.3.7 - 31 Jul 2025
 Bug fix:
+- [#1314](https://github.com/elysiajs/elysia/issues/1314) coerce TransformDecodeError to ValidationError
 - [#1313](https://github.com/elysiajs/elysia/pull/1313) onRequest not firing
 
 # 1.3.6 - 24 Jul 2025